ADEN-12036 - Working on moving PPID initialisation
